Refactor DType definition into dtypes_core.py (#468)
Summary:
Pull Request resolved: https://github.com/pytorch/torcharrow/pull/468

dtypes.py contains of two parts:
(1) Standard Arrow-compatible DataType definition (Int8/16/32/64, Float32/64, String, List, Map, Struct)
(2) Utility functions around DType

The DType definition is quite standalone and stable; Refactor DataType definition into seperate file so allows reuse in next gen TorchArrow, such as TorchArrow-UPM or Tensor-based TA

In theory we could let dtypes.py only contain DType definition, and move other things into `dtype_util.py`. Starting with this first step.

Reviewed By: dracifer

Differential Revision: D38358389

fbshipit-source-id: 632037498290222a60794e26d0db0c98c73f1391
